+++ src/Agent/NewRelic/Agent/Core/Attributes/AttributeDefinitionService.cs
@@ -133,6 +133,17 @@ namespace NewRelic.Agent.Core.Attributes
         AttributeDefinition<TypeAttributeValue, string> GetTypeAttribute(TypeAttributeValue destination);
 
         AttributeDefinition<bool, bool> LlmTransaction { get; }
+
+        AttributeDefinition<string, string> CloudAccountId { get; }
+        AttributeDefinition<string, string> CloudRegion { get; }
+         AttributeDefinition<string, string> MessagingSystemName { get; }
+        AttributeDefinition<string, string> MessagingDestinationName { get; }
+        AttributeDefinition<string, string> BrokerServerAddress { get; }
+        AttributeDefinition<int, int> BrokerServerPort { get; }
+        AttributeDefinition<string, string> MessageQueueName { get; }
+        AttributeDefinition<string, string> MessageRoutingKey { get; }
+        AttributeDefinition<string, string> MessagingRabbitMqDestinationRoutingKey { get; }
+        AttributeDefinition<string, string> MessagingDestinationPublishName { get; }
     }
 
 
@@ -177,7 +188,7 @@ namespace NewRelic.Agent.Core.Attributes
         private readonly ConcurrentDictionary<string, AttributeDefinition<string, string>> _requestParameterAttributes = new ConcurrentDictionary<string, AttributeDefinition<string, string>>();
         private readonly ConcurrentDictionary<string, AttributeDefinition<string, string>> _requestHeadersAttributes = new ConcurrentDictionary<string, AttributeDefinition<string, string>>();
         private readonly ConcurrentDictionary<string, AttributeDefinition<object, object>> _lambdaAttributes = new ConcurrentDictionary<string, AttributeDefinition<object, object>>();
-        
+
         private readonly ConcurrentDictionary<TypeAttributeValue, AttributeDefinition<TypeAttributeValue, string>> _typeAttributes = new ConcurrentDictionary<TypeAttributeValue, AttributeDefinition<TypeAttributeValue, string>>();
 
 
@@ -1078,6 +1089,74 @@ namespace NewRelic.Agent.Core.Attributes
             AttributeDefinitionBuilder.CreateBool("llm", AttributeClassification.AgentAttributes)
                 .AppliesTo(AttributeDestinations.TransactionEvent)
                 .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _cloudAccountId;
+        public AttributeDefinition<string, string> CloudAccountId => _cloudAccountId ?? (_cloudAccountId =
+            AttributeDefinitionBuilder.CreateString("cloud.account.id", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _cloudRegion;
+        public AttributeDefinition<string, string> CloudRegion => _cloudRegion ?? (_cloudRegion =
+            AttributeDefinitionBuilder.CreateString("cloud.region", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _messagingSystem;
+        public AttributeDefinition<string, string> MessagingSystemName => _messagingSystem ?? (_messagingSystem =
+            AttributeDefinitionBuilder.CreateString("messaging.system", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _messagingDestinationName;
+        public AttributeDefinition<string, string> MessagingDestinationName => _messagingDestinationName ?? (_messagingDestinationName =
+            AttributeDefinitionBuilder.CreateString("messaging.destination.name", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+
+        // new attribute for MessageBrokers - same name as the Externals attribute, but in Agent attributes.
+        // From messaging api spec.
+        private AttributeDefinition<string, string> _brokerServerAddress;
+        public AttributeDefinition<string, string> BrokerServerAddress => _brokerServerAddress ?? (_brokerServerAddress =
+            AttributeDefinitionBuilder.CreateString("server.address", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<int, int> _brokerServerPort;
+        public AttributeDefinition<int, int> BrokerServerPort => _brokerServerPort ?? (_brokerServerPort =
+            AttributeDefinitionBuilder.CreateInt("server.port", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .AppliesTo(AttributeDestinations.TransactionTrace)
+                .Build(_attribFilter));
+        
+        private AttributeDefinition<string, string> _messageQueueName;
+        public AttributeDefinition<string, string> MessageQueueName => _messageQueueName ?? (_messageQueueName =
+            AttributeDefinitionBuilder.CreateString("message.queueName", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.All)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _messageRoutingKey;
+        public AttributeDefinition<string, string> MessageRoutingKey => _messageRoutingKey ?? (_messageRoutingKey =
+            AttributeDefinitionBuilder.CreateString("message.routingKey", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.All)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _messagingRabbitMqDestinationRoutingKey;
+        public AttributeDefinition<string, string> MessagingRabbitMqDestinationRoutingKey => _messagingRabbitMqDestinationRoutingKey ?? (_messagingRabbitMqDestinationRoutingKey =
+            AttributeDefinitionBuilder.CreateString("messaging.rabbitmq.destination.routing_key", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
+                .Build(_attribFilter));
+
+        private AttributeDefinition<string, string> _messagingDestinationPublishName;
+        public AttributeDefinition<string, string> MessagingDestinationPublishName => _messagingDestinationPublishName ?? (_messagingDestinationPublishName =
+            AttributeDefinitionBuilder.CreateString("messaging.destination_publish.name", AttributeClassification.AgentAttributes)
+                .AppliesTo(AttributeDestinations.SpanEvent)
                 .Build(_attribFilter));
     }
 }