+++ src/Agent/NewRelic/Agent/Core/Transactions/NoOpTransaction.cs
@@ -73,7 +73,10 @@ namespace NewRelic.Agent.Core.Transactions
             return Segment.NoOpSegment;
         }
 
-        public ISegment StartMessageBrokerSegment(MethodCall methodCall, MessageBrokerDestinationType destinationType, MessageBrokerAction operation, string brokerVendorName, string destinationName)
+        public ISegment StartMessageBrokerSegment(MethodCall methodCall, MessageBrokerDestinationType destinationType,
+            MessageBrokerAction operation, string brokerVendorName, string destinationName,
+            string messagingSystemName = null, string cloudAccountId = null, string cloudRegion = null,
+            string serverAddress = null, int? serverPort = null, string routingKey = null)
         {
 #if DEBUG
             Log.Finest("Skipping StartMessageBrokerSegment outside of a transaction");