+++ src/Agent/NewRelic/Agent/Extensions/NewRelic.Agent.Extensions/AwsSdk/SqsHelper.cs
@@ -18,6 +18,7 @@ namespace NewRelic.Agent.Extensions.AwsSdk
         private static ConcurrentDictionary<Type, Func<object, IDictionary>> _getMessageAttributes = new();
         private static Func<object> _messageAttributeValueTypeFactory;
 
+        public const string MessagingSystemName = "aws_sqs";
         public const string VendorName = "SQS";
 
         private class SqsAttributes
@@ -25,6 +26,7 @@ namespace NewRelic.Agent.Extensions.AwsSdk
             public string QueueName { get; }
             public string CloudId { get; }
             public string Region { get; }
+            public string ServerAddress { get; }
 
             // https://sqs.us-east-2.amazonaws.com/123456789012/MyQueue
             public SqsAttributes(string url)
@@ -51,12 +53,16 @@ namespace NewRelic.Agent.Extensions.AwsSdk
 
                 // subdomain[0] should always be "sqs"
                 Region = subdomain[1];
+
+                ServerAddress = new Uri(url).Host;
             }
         }
+
         public static ISegment GenerateSegment(ITransaction transaction, MethodCall methodCall, string url, MessageBrokerAction action)
         {
             var attr = new SqsAttributes(url);
-            var segment = transaction.StartMessageBrokerSegment(methodCall, MessageBrokerDestinationType.Queue, action, VendorName, attr.QueueName);
+
+            var segment = transaction.StartMessageBrokerSegment(methodCall, MessageBrokerDestinationType.Queue, action, VendorName, destinationName: attr.QueueName, messagingSystemName: MessagingSystemName, cloudAccountId: attr.CloudId, cloudRegion: attr.Region);
             segment.GetExperimentalApi().MakeLeaf();
 
             return segment;
@@ -71,7 +77,7 @@ namespace NewRelic.Agent.Extensions.AwsSdk
 
             var setHeaders = new Action<object, string, string>((smr, key, value) =>
             {
-                var getMessageAttributes = _getMessageAttributes.GetOrAdd(smr.GetType(), t   => VisibilityBypasser.Instance.GeneratePropertyAccessor<IDictionary>(t, "MessageAttributes"));
+                var getMessageAttributes = _getMessageAttributes.GetOrAdd(smr.GetType(), t => VisibilityBypasser.Instance.GeneratePropertyAccessor<IDictionary>(t, "MessageAttributes"));
                 var messageAttributes = getMessageAttributes(smr);
 
                 // if we can't add all DT headers, don't add any