+++ src/Agent/NewRelic/Agent/Extensions/NewRelic.Agent.Extensions/Parsing/SqlParser.cs
@@ -239,7 +239,8 @@ namespace NewRelic.Agent.Extensions.Parsing
                     }
                 }
 
-                return _nullParsedStatementStore.GetOrAdd(datastoreVendor, x => new ParsedSqlStatement(datastoreVendor, null, null));
+                // return a null statement with the model name if found, otherwise cache and return a null statement with no model name
+                return !string.IsNullOrEmpty(explicitModel) ? new ParsedSqlStatement(datastoreVendor, explicitModel, null) : _nullParsedStatementStore.GetOrAdd(datastoreVendor, x => new ParsedSqlStatement(datastoreVendor, null, null));
             };
         }
 
@@ -295,7 +296,7 @@ namespace NewRelic.Agent.Extensions.Parsing
                         model = "ParseError";
                     }
                 }
-                
+
                 return CreateParsedDatabaseStatement(vendor, GetFullModel(model, explicitModel));
             }
 
@@ -323,7 +324,7 @@ namespace NewRelic.Agent.Extensions.Parsing
                     string model = FromMatcher.Match(statement).Success
                         ? "(subquery)"
                         : "VARIABLE";
-                    
+
                     return new ParsedSqlStatement(vendor, GetFullModel(model, explicitModel), "select");
                 }
                 return null;
@@ -414,7 +415,7 @@ namespace NewRelic.Agent.Extensions.Parsing
                     Log.Debug("Not executing explain plan since DbType is Object.");
                     return false;
                 }
-                
+
                 // Parameter names can be supplied with the prefix @ or without
                 // if is supplied, remove the @ to the beginning of the param name
                 // This is to deal with regex issues, see below